helper method was moved to appropriate file.
