+++ clusterman/kubernetes/util.py
@@ -162,18 +162,3 @@ def selector_term_matches_requirement(
                 if match_expression == selector_requirement:
                     return True
     return False
-
-
-def strtobool(val: str) -> int:
-    """Convert a string representation of truth to true (1) or false (0).
-    True values are 'y', 'yes', 't', 'true', 'on', and '1'; false values
-    are 'n', 'no', 'f', 'false', 'off', and '0'.  Raises ValueError if
-    'val' is anything else.
-    """
-    val = val.lower()
-    if val in ("y", "yes", "t", "true", "on", "1"):
-        return 1
-    elif val in ("n", "no", "f", "false", "off", "0"):
-        return 0
-    else:
-        raise ValueError("invalid truth value %r" % (val,))
